backend/requirements.txt
Flask==2.2.2
gunicorn==20.1.0
firebase-admin==6.1.0
google-generativeai==0.3.1
python-dotenv==0.21.0
flask-cors==4.0.0
